# openapi.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8081","description":"Generated server url"}],"tags":[{"name":"Customers","description":"Customers API"}],"paths":{"/customers/{id}":{"get":{"tags":["Customers"],"summary":"Get a Customer","operationId":"getCustomerById","parameters":[{"name":"id","in":"path","description":"The ID of the customer to retrieve","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"Customer retrieved","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CustomerDTO"}}}},"default":{"description":"Unexpected Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Error"}}}}}},"put":{"tags":["Customers"],"summary":"Update a customer","operationId":"updateCustomer","parameters":[{"name":"id","in":"path","description":"The ID of the customer to update","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CustomerDTO"}}},"required":true},"responses":{"200":{"description":"Customer updated","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CustomerDTO"}}}},"default":{"description":"Unexpected Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Error"}}}}}},"delete":{"tags":["Customers"],"summary":"Delete a customer","operationId":"deleteCustomer","parameters":[{"name":"id","in":"path","description":"The ID of the customer to delete","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"Customer deleted"},"default":{"description":"Unexpected Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Error"}}}}}}},"/customers":{"get":{"tags":["Customers"],"summary":"Get all customers","operationId":"getCustomers","responses":{"200":{"description":"Success","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CustomerDTO"}}}}},"default":{"description":"Unexpected Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Error"}}}}}},"post":{"tags":["Customers"],"summary":"Create a Customer","operationId":"createCustomer","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CustomerDTO"}}},"required":true},"responses":{"201":{"description":"Customer created","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CustomerDTO"}}}},"default":{"description":"Unexpected Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Error"}}}}}}}},"components":{"schemas":{"CustomerDTO":{"type":"object","properties":{"customerId":{"type":"integer","format":"int64"},"name":{"type":"string"},"city":{"type":"string"}}},"Error":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"message":{"type":"string"}}}}}}