# services/customer-service/src/main/java/de/ostfalia/serp/customer/api/CustomersApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.2.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package de.ostfalia.serp.customer.api;

import de.ostfalia.serp.customer.dto.CustomerDTO;
import de.ostfalia.serp.customer.dto.Error;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-03-14T15:15:19.485031803+01:00[Europe/Berlin]")
@Validated
@Tag(name = "Customers", description = "Customers API")
public interface CustomersApi {

    /**
     * POST /customers : Create a Customer
     *
     * @param customerDTO  (required)
     * @return Customer created (status code 201)
     *         or Unexpected Error (status code 200)
     */
    @Operation(
        operationId = "createCustomer",
        summary = "Create a Customer",
        tags = { "Customers" },
        responses = {
            @ApiResponse(responseCode = "201", description = "Customer created", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = CustomerDTO.class))
            }),
            @ApiResponse(responseCode = "default", description = "Unexpected Error", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Error.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/customers",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    ResponseEntity<CustomerDTO> createCustomer(
        @Parameter(name = "CustomerDTO", description = "", required = true) @Valid @RequestBody CustomerDTO customerDTO
    );


    /**
     * DELETE /customers/{id} : Delete a customer
     *
     * @param id The ID of the customer to delete (required)
     * @return Customer deleted (status code 200)
     *         or Unexpected Error (status code 200)
     */
    @Operation(
        operationId = "deleteCustomer",
        summary = "Delete a customer",
        tags = { "Customers" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Customer deleted"),
            @ApiResponse(responseCode = "default", description = "Unexpected Error", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Error.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = "/customers/{id}",
        produces = { "application/json" }
    )
    
    ResponseEntity<Void> deleteCustomer(
        @Parameter(name = "id", description = "The ID of the customer to delete", required = true, in = ParameterIn.PATH) @PathVariable("id") Long id
    );


    /**
     * GET /customers/{id} : Get a Customer
     *
     * @param id The ID of the customer to retrieve (required)
     * @return Customer retrieved (status code 200)
     *         or Unexpected Error (status code 200)
     */
    @Operation(
        operationId = "getCustomerById",
        summary = "Get a Customer",
        tags = { "Customers" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Customer retrieved", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = CustomerDTO.class))
            }),
            @ApiResponse(responseCode = "default", description = "Unexpected Error", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Error.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/customers/{id}",
        produces = { "application/json" }
    )
    
    ResponseEntity<CustomerDTO> getCustomerById(
        @Parameter(name = "id", description = "The ID of the customer to retrieve", required = true, in = ParameterIn.PATH) @PathVariable("id") Long id
    );


    /**
     * GET /customers : Get all customers
     *
     * @return Success (status code 200)
     *         or Unexpected Error (status code 200)
     */
    @Operation(
        operationId = "getCustomers",
        summary = "Get all customers",
        tags = { "Customers" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Success", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = CustomerDTO.class)))
            }),
            @ApiResponse(responseCode = "default", description = "Unexpected Error", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Error.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/customers",
        produces = { "application/json" }
    )
    
    ResponseEntity<List<CustomerDTO>> getCustomers(
        
    );


    /**
     * PUT /customers/{id} : Update a customer
     *
     * @param id The ID of the customer to update (required)
     * @param customerDTO  (required)
     * @return Customer updated (status code 200)
     *         or Unexpected Error (status code 200)
     */
    @Operation(
        operationId = "updateCustomer",
        summary = "Update a customer",
        tags = { "Customers" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Customer updated", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = CustomerDTO.class))
            }),
            @ApiResponse(responseCode = "default", description = "Unexpected Error", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Error.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = "/customers/{id}",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    ResponseEntity<CustomerDTO> updateCustomer(
        @Parameter(name = "id", description = "The ID of the customer to update", required = true, in = ParameterIn.PATH) @PathVariable("id") Long id,
        @Parameter(name = "CustomerDTO", description = "", required = true) @Valid @RequestBody CustomerDTO customerDTO
    );

}